requirements.txt
awswrangler==2.7.0
python-dateutil==2.8.1
tqdm==4.60.0
pandas==1.2.4
numpy==1.20.3
urllib3==1.25.11
pytest==7.4.2
pytest-sugar==0.9.7
